requirements.txt
google-genai
beautifulsoup4
selenium
undetected-chromedriver
python-dotenv
pymupdf
requests
